move createView
